move private funcs to factory
